backend/pyproject.toml
[project]
name = "podi-backend"
version = "0.1.0"
description = "PODI AI 后端服务"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.25",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "oss2>=2.19",
  "aliyun-python-sdk-sts>=3.1",
  "alembic>=1.13",
  "celery>=5.3",
  "boto3>=1.34",
  "python-multipart>=0.0.9",
  "httpx>=0.26",
  "passlib[bcrypt]>=1.7",
  "pyjwt[crypto]>=2.8",
  "redis>=5.0",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "coverage>=7.4",
  "black>=24.3",
  "ruff>=0.6"
]

[tool.setuptools]
packages = ["app"]
